Previously, every source file contained a line like one of the following at the top to control the compiler's optimization policy:
  (declaim #.*optimize-default*)
  (declaim #.*optimize-fast-unsafe*)

This change implements a new way of doing roughly the same thing.
Now, in your ASDF system definitions, use one of the following for
Lisp source files:

  (:file "foobar")
  (:fast-unsafe-source-file "foobar")

The first form compiles foobar with the current compiler optimization
settings.  The second form compiles it with the policy stored in
com.google.base:*optimize-fast-unsafe*, which specifies high
optimization.